--- src/main/java/org/sugar_square/community_service/repository/CategoryRepository.java
@@ -1,9 +1,0 @@
-package org.sugar_square.community_service.repository;
-
-import org.springframework.stereotype.Repository;
-import org.sugar_square.community_service.domain.board.Category;
-
-@Repository
-public interface CategoryRepository extends BaseRepository<Category, Long>{
-
-}
