--- src/main/java/org/sugar_square/community_service/repository/CommentRepository.java
@@ -1,9 +1,0 @@
-package org.sugar_square.community_service.repository;
-
-import org.springframework.stereotype.Repository;
-import org.sugar_square.community_service.domain.board.Comment;
-
-@Repository
-public interface CommentRepository extends BaseRepository<Comment, Long>{
-
-}
