--- src/main/java/org/sugar_square/community_service/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package org.sugar_square.community_service.repository;
-
-import org.springframework.stereotype.Repository;
-import org.sugar_square.community_service.domain.member.Member;
-
-@Repository
-public interface MemberRepository extends BaseRepository<Member, Long>{
-
-}
